[-Wformat] format specifies type 'int' but the argument has type 'size_type' (aka 'unsigned long')
GraySkeletonCPP/InteractiveSkeletonizer.h#L103 [-Wformat] format specifies type 'int' but the argument has type 'size_type' (aka 'unsigned long')
ProteinMorph/SSECorrespondenceFinder.h#L605 [-Wformat] format specifies type 'int' but the argument has type 'size_type' (aka 'unsigned long')
ProteinMorph/TriangleMesh.h#L150 [-Wformat] format specifies type 'int' but the argument has type 'size_type' (aka 'unsigned long')

+++ ProteinMorph/SSECorrespondenceFinder.h
@@ -602,7 +602,7 @@ namespace wustl_mm {
 			
 			tm.PushCurrentTime();
 			vector< set<unsigned long long> > allCliques = parentGraph.GetAllCliquesTriangleApprox(smallestCliqueSize);
-			printf("(*%d cliques found ", allCliques.size());
+			printf("(*%d cliques found ", (int)allCliques.size());
 			tm.PopAndDisplayTime("%f seconds!*)\n");
 
 			tm.PushCurrentTime();
